adjust graph and test
