--- Race Element.Graph/Node/ValueNodes/AbstractFloatNode.cs
@@ -1,8 +1,0 @@
-﻿namespace RaceElement.Graph.Node.ValueNodes;
-
-public abstract record class AbstractFloatNode : AbstractNode
-{
-    public AbstractFloatNode(float value) => Value = value;
-
-    public float Value { get; init; }
-}
